Renamed StateEventApplier to AggregateState as its better communicates what it actually is
