+++ Source/EventFlow.Tests/EventFlow.Tests.csproj
@@ -72,7 +72,7 @@
     <Compile Include="IntegrationTests\InMemoryConfiguration.cs" />
     <Compile Include="IntegrationTests\ReadStores\InMemoryReadModelStoreTests.cs" />
     <Compile Include="UnitTests\Aggregates\AggregateRootApplyEventTests.cs" />
-    <Compile Include="UnitTests\Aggregates\EventAppliers\StateEventApplierTests.cs" />
+    <Compile Include="UnitTests\Aggregates\AggregateStateTests.cs" />
     <Compile Include="UnitTests\Queries\QueryProcessorTests.cs" />
     <Compile Include="UnitTests\ReadStores\ReadModelDomainEventApplierTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
